+++ ui/src/layouts/chats/presentation/chatbar/coroutines.rs
@@ -1,18 +1,21 @@
-use std::path::PathBuf;
+use std::{
+    path::PathBuf,
+    time::{Duration, Instant},
+};
 
 use common::{
     state::{Action, State},
     warp_runner::{RayGunCmd, WarpCmd},
-    WARP_CMD_CH,
+    STATIC_ARGS, WARP_CMD_CH,
 };
 use dioxus::prelude::*;
 use futures::{channel::oneshot, StreamExt};
 use uuid::Uuid;
-use warp::raygun::Location;
+use warp::raygun::{self, Location};
 
 use crate::layouts::chats::data::{self, ChatProps, DEFAULT_MESSAGES_TO_TAKE};
 
-use super::ChatInput;
+use super::{ChatInput, TypingIndicator, TypingInfo};
 
 pub fn get_msg_ch<'a>(
     cx: &Scoped<'a, ChatProps>,
@@ -119,6 +122,84 @@ pub fn get_scroll_ch<'a>(
                         );
                     }
                     Err(e) => log::error!("{e}"),
+                }
+            }
+        }
+    })
+    .clone()
+}
+
+// typing indicator notes
+// consider side A, the local side, and side B, the remote side
+// side A -> (typing indicator) -> side B
+// side B removes the typing indicator after a timeout
+// side A doesn't want to send too many typing indicators, say once every 4-5 seconds
+// should we consider matching the timeout with the send frequency so we can closely match if a person is straight up typing for 5 mins straight.
+
+// tracks if the local participant is typing
+// re-sends typing indicator in response to the Refresh command
+pub fn get_typing_ch<'a>(cx: &Scoped<'a, ChatProps>) -> Coroutine<TypingIndicator> {
+    use_coroutine(cx, |mut rx: UnboundedReceiver<TypingIndicator>| {
+        // to_owned![];
+        async move {
+            let mut typing_info: Option<TypingInfo> = None;
+            let warp_cmd_tx = WARP_CMD_CH.tx.clone();
+
+            let send_typing_indicator = |conv_id| async move {
+                let (tx, rx) = oneshot::channel::<Result<(), warp::error::Error>>();
+                let event = raygun::MessageEvent::Typing;
+                if let Err(e) = warp_cmd_tx.send(WarpCmd::RayGun(RayGunCmd::SendEvent {
+                    conv_id,
+                    event,
+                    rsp: tx,
+                })) {
+                    log::error!("failed to send warp command: {}", e);
+                    // return from the closure
+                    return;
+                }
+                let rsp = rx.await.expect("command canceled");
+                if let Err(e) = rsp {
+                    log::error!("failed to send typing indicator: {}", e);
+                }
+            };
+
+            while let Some(indicator) = rx.next().await {
+                match indicator {
+                    TypingIndicator::Typing(chat_id) => {
+                        // if typing_info was none or the chat id changed, send the indicator immediately
+                        let should_send_indicator = match typing_info {
+                            None => true,
+                            Some(info) => info.chat_id != chat_id,
+                        };
+                        if should_send_indicator {
+                            send_typing_indicator.clone()(chat_id).await;
+                        }
+                        typing_info = Some(TypingInfo {
+                            chat_id,
+                            last_update: Instant::now(),
+                        });
+                    }
+                    TypingIndicator::NotTyping => {
+                        typing_info = None;
+                    }
+                    TypingIndicator::Refresh(conv_id) => {
+                        let info = match &typing_info {
+                            Some(i) => i.clone(),
+                            None => continue,
+                        };
+                        if info.chat_id != conv_id {
+                            typing_info = None;
+                            continue;
+                        }
+                        // todo: verify duration for timeout
+                        let now = Instant::now();
+                        if now - info.last_update
+                            <= (Duration::from_secs(STATIC_ARGS.typing_indicator_timeout)
+                                - Duration::from_millis(500))
+                        {
+                            send_typing_indicator.clone()(conv_id).await;
+                        }
+                    }
                 }
             }
         }
